requirements.txt
boto3==1.35.99
botocore==1.35.99
dnspython==2.6.1
et-xmlfile==2.0.0
jmespath==1.0.1
numpy==1.24.4
pandas==2.0.3
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
s3transfer==0.10.4
six==1.17.0
tzdata==2024.2
urllib3==1.26.20
XlsxWriter==3.2.0
